+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Make Ctrl-C cancellation instantaneous, no more waiting for the jobs to finish
 - Use Rocky Linux for simple RPM builds
 - Add alpine APK target [#75](https://github.com/vv9k/pkger/pull/75)
+- Add `print-completions` subcommand that prints shell completions to stdout. Supported shells are: bash, zsh, fish, powershell, elvish
 
 # 0.6.0
 - Use default OS config directory instead of home directory to read the configuration file by default.